# schemas/src/swagger-schema.json
{"openapi":"3.0.0","paths":{"/v1/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/v1/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationBody"}}}},"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/v1/api/auth":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginBody"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"allOf":[{"title":"SingleResult","$ref":"#/components/schemas/SingleResult"},{"properties":{"data":{"$ref":"#/components/schemas/User"}}}]}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/v1/api/auth/session/refresh":{"post":{"operationId":"AuthController_refreshAccessToken","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/v1/api/users/me":{"get":{"operationId":"UserController_getMe","parameters":[],"responses":{"200":{"description":"Get user successfully"}},"tags":["Users"]}},"/v1/api/users/user":{"get":{"operationId":"UserController_findUser","parameters":[{"name":"username","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get user successfully"}},"tags":["Users"]}},"/v1/api/settings/me":{"get":{"operationId":"SettingController_getMe","parameters":[],"responses":{"200":{"description":"Get user successfully"}},"tags":["Settings"]}},"/v1/api/friends/requests":{"get":{"operationId":"FriendController_getListFriendRequest","parameters":[],"responses":{"200":{"description":"Get paginationresult successfully"}},"tags":["Friends"]}},"/v1/api/friends/request/{id}":{"put":{"operationId":"FriendController_acceptRequest","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendRequestUpdatingBody"}}}},"responses":{"200":{"description":"Update paginationresult successfully","content":{"application/json":{"schema":{"allOf":[{"title":"SingleResult","$ref":"#/components/schemas/SingleResult"},{"properties":{"data":{"$ref":"#/components/schemas/PaginationResult"}}}]}}}}},"tags":["Friends"]}},"/v1/api/friends":{"get":{"operationId":"FriendController_getListFriend","parameters":[],"responses":{"200":{"description":"Get paginationresult successfully"}},"tags":["Friends"]}}},"info":{"title":"api_base","description":"Bidding API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserCreationBody":{"type":"object","properties":{"firstName":{"type":"string","default":"John"},"lastName":{"type":"string","default":"Doe"},"username":{"type":"string"},"password":{"type":"string","default":"Bidding@123"},"role":{"type":"string","enum":["admin","user"]}},"required":["firstName","lastName","username","password","role"]},"SingleResult":{"type":"object","properties":{"data":{"type":"object"}},"required":["data"]},"User":{"type":"object","properties":{"id":{"type":"number"},"isArchived":{"type":"boolean"},"dateCreate":{"format":"date-time","type":"string"},"dateUpdate":{"format":"date-time","type":"string"},"username":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string","enum":["admin","user"]}},"required":["id","username","password","firstName","lastName","role"]},"UserLoginBody":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"PaginationResult":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array","items":{"type":"object"}}},"total":{"type":"number","example":100,"description":"Total number of items"},"page":{"type":"number","example":1,"description":"Current page number"},"limit":{"type":"number","example":10,"description":"Number of items per page"}},"required":["data","total","page","limit"]},"FriendRequestUpdatingBody":{"type":"object","properties":{"requestId":{"type":"number"},"senderId":{"type":"number"},"receiverId":{"type":"number"},"status":{"type":"string","enum":["Pending","Accepted","Rejected"]}},"required":["requestId","senderId","receiverId","status"]}}}}